Keycloak_app/requirements.txt
Django==3.2.13
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
mysqlclient==2.1.1
python-keycloak
requests